+++ chai/info.h
@@ -36,7 +36,7 @@ void printMoveStatus(board_t* b, move_t move);
  *
  * @returns The move in algebraic notation.
  */
-std::string getStringMove(board_t* b, const int move);
+std::string getStringMove(board_t* b, const move_t move);
 
 ///**
 // * Print all flags and attributes of given move.
@@ -81,13 +81,6 @@ void printPvLine(board_t* b, move_t* pvLine, int d, int score);
  * @param  logMsg Message to write into log file.
  */
 void log(std::string logMsg);
-
-/**
- * Gets the time.
- *
- * @returns The time.
- */
-std::string getTime();
 
 /**
  * Read input from command line during search. Sets stop to true,